requirements.txt
opencv-python-headless==4.10.0.84
gradio==5.0
huggingface-hub==0.26.2
python-dotenv==1.0.1
paddleocr==2.9.1
paddlepaddle-gpu==2.5.2
ultralytics==8.3.36
fastapi==0.115.5
pillow==10.2.0
uvicorn==0.27.1
transformers==4.35.2